Add 'other' new special events to eventSheet's context menu
This also moves other events menu array out of the toolbar class and into the eventSheet class. The toolbar class still gets it as a prop and still works as before
